requirements.txt
simplelogger-python==0.1.10
py-config-handler==1.0.2
pandas==1.4.2
pywin32==303
